fix: reduce file line counts to comply with 1500-line CI limit
- Remove a comment in github-merge.lib.mjs to bring it from 1503 to 1500 lines
- Extract Issue #1588 tests from test-merge-queue.mjs (1599 lines) into
  separate test-merge-queue-cancel-1588.mjs file (1493 + 154 lines)
